gnotravex: make better use of GtkApplication
One feature of GtkApplication is application uniqueness.  This means
that when the application is started for the second time, the first one
is focused rather than creating a second window.

For this reason, we should avoid doing all of the setup work for the
application until we are sure that we are going to be the first
instance.  That's what the startup() virtual function is for.

The activate() virtual function runs each time we are invoked (for the
first time, or also in the future).  We can use this to present the
window.

Finally, the newly-introduced shutdown() function does the opposite of
startup().